feat: sidebar prompt suggestion (#203)
Branched off of #202 but this is functionally separate and should be
merged after

This adds a basic connection between the sidebar prompt builder and the
`/chat` api endpoint to generate some query prompts. Right now, this
will only work when there is no active chat (as defined by there being
no messages in the UI state).

There is some light debouncing that happens to account for users
clicking around.


https://github.com/user-attachments/assets/95437fac-68f8-4a6b-b926-26671e2a9b35

This doesn't address this, but this feature does make it _feel_ like we
need to drop the sidebar context if a generated example prompt is used
-- one would think that an LLM generated query would seed the chat w/
enough requisite context. But that's mostly feeling based and for
another time/PR